+++ web/v2/backend/src/main/java/com/tourneygen/web/Services/EloService.java
@@ -1,6 +1,6 @@
 package com.tourneygen.web.Services;
 
 public class EloService {
-  public static final float ELO_INITIAL_VALUE = 800.0f;
-  public static final float ELO_SCALE_CONST = 30.0f;
+  public static final int ELO_INITIAL_VALUE = 800;
+  public static final int ELO_SCALE_CONST = 30;
 }
